Merge CSS into JS files to reduce load time
When an asset file under `https://unpkg.com/foremark/` is requested,
UNPKG redirects a browser to the URL with a specific version number
(e.g., `https://unpkg.com/foremark@0.0.4/`), essentially doubling the
load latency. This makes it crucial to reduce the length of a load
chain.

This commit modifies the webpack build process to inline CSS into the
generated JS chunks, reducing the length of a load chain by one.
This was accomplished through the use of `style-loader`, which
automatically injects CSS to the DOM by generating a `<style>` tag at
runtime.