Refactor `TableOfContents`
`Node` (representing a TOC entry) is separated into two layers of
data models: `Node` used as the data source as well as for
identification, and `ViewNode` as a view model storing the state of
displayed TOC entries.

This makes it possible to have multiple views of a single TOC tree while
remembering the state of each view.