fix(toc): do not set `href` for disabled items
`href=""` links to the current page. Although it can be prevented by
listening to `click` events, this solution doesn't work when JavaScript
is disabled and displaying a server-side rendered page. The same thing
happens during JavaScript source files are being loaded, which is
pretty annoying because the loading of source files starts over every
time the visitor clicks a disabled TOC entry, so they won't be able to
expand a collapsed, disabled TOC entry unless they know they have to
wait until JavaScript source files are loaded.

A hyperlink is completely disabled whether JavaScript is enable or not
if `href` is not set.